! ip needs to be bigint since Postgres doesn't support unsigned ints

+++ other/install_2-1_postgresql.sql
@@ -1243,7 +1243,7 @@ CREATE TABLE {$db_prefix}log_online (
   log_time int NOT NULL default '0',
   id_member int NOT NULL default '0',
   id_spider smallint NOT NULL default '0',
-  ip int NOT NULL default '0',
+  ip bigint NOT NULL default '0',
   url text NOT NULL,
   PRIMARY KEY (session)
 );
